IPUMS code and data
Added the IPUMS code to create the final proportions csv. This includes the R code in order to read in the .dat file from IPUMS and creata a csv from the .dat file. Python code that reads in the csv and creates another csv with just the proportions. And lastly, a python file specific for creating visualizations from the proportions csv.